requirements.txt
annotated-types==0.7.0
anyio==4.4.0
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
distro==1.9.0
google-api-core==2.19.0
google-auth==2.30.0
google-cloud-documentai==2.27.0
googleapis-common-protos==1.63.1
grpcio==1.64.1
grpcio-status==1.62.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
loguru==0.7.2
openai==1.34.0
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.7.4
pydantic_core==2.18.4
pypdf==4.2.0
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.32.3
rsa==4.9
sniffio==1.3.1
tqdm==4.66.4
typing_extensions==4.12.2
urllib3==2.2.1
